+++ tests/testthat/test-cbbinom.R
@@ -7,6 +7,7 @@ test_alpha <- 2
 test_beta <- 4
 test_delta <- 1e-6
 test_tol <- NULL
+test_prec <- 20
 test_root_tol <- 1e-6
 test_max_iter <- 10000L
 
@@ -23,8 +24,7 @@ testthat::test_that(
                 alpha = test_alpha, beta = test_beta, log.p = TRUE) -
          pcbbinom(q = test_x - test_delta, size = test_size,
                   alpha = test_alpha, beta = test_beta, log.p = TRUE)) /
-        (2 * test_delta) * test_val,
-      tolerance = 1e-6
+        (2 * test_delta) * test_val
     )
   }
 )
@@ -55,7 +55,7 @@ testthat::test_that(
       cpp_qcbbinom(p = test_val_log, size = test_size,
                    alpha = test_alpha, beta = test_beta,
                    lower_tail = TRUE, log_p = TRUE,
-                   p_tol = test_tol, p_max_iter = test_max_iter,
+                   p_tol = test_tol, p_max_iter = test_max_iter, p_prec = test_prec,
                    root_tol = test_root_tol, root_max_iter = test_max_iter),
       test_x
     )
@@ -89,6 +89,7 @@ testthat::test_that(
                    x = 1,
                    tol = test_tol,
                    max_iter = test_max_iter,
+                   prec = test_prec,
                    check_mode = TRUE,
                    log = FALSE),
       101/5460